Moved function send_newsletter
